requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.4
opencv-python==4.11.0.86
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pycparser==2.22
pyOpenSSL==25.0.0
PySocks==1.7.1
pytesseract==0.3.13
pytest==8.3.5
PyYAML==6.0.2
requests==2.32.3
selenium==4.15.1
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.2.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==1.26.20
urllib3-secure-extra==0.1.0
webdriver-manager==4.0.1
websocket-client==1.8.0
wsproto==1.2.0
